Csutomer Order Project
I made a product sales back-end application using Java and Spring Boot. I created the users and products with postman and used PostgreSQL as the database. According to the existence of the entered user in the database, the system recommends the products in the system to the selected user. The user creates an order by selecting the product they want. As soon as the order is created, the order record is automatically triggered to the database. While doing this project, I used Spring JPA, MapStruct and Spring  WebClient frequently, and I designed the project according to the MVC architecture and ensured that it was in an order.